src/requirements.txt
pandas
numpy
opencv_python
Pillow
matplotlib
geopy
python-dotenv
black
